Address some review
* Add docs
* Rename `ShardManagerMessage::Err` to `Quit`
* Move `ShardQueue` into its own file
* Change runner_info `Mutex` to a sync implementation and don't block
  when locking it
* Outline the shutdown logic in the `Drop` impl for `ShardManager`
